+++ src/stats_duck_extension.cpp
@@ -11,6 +11,7 @@
 #include "sign_test_function.hpp"
 #include "adjust_p_function.hpp"
 #include "poibin_function.hpp"
+#include "auto_bin_function.hpp"
 #include "table_one_function.hpp"
 #include "normality_function.hpp"
 #include "ks_test_function.hpp"
@@ -55,6 +56,10 @@ static void LoadInternal(ExtensionLoader &loader) {
 
 	// Poisson Binomial CDF
 	RegisterPoibinCdf(loader);
+
+	// Auto-binning helpers
+	RegisterBinEdges(loader);
+	RegisterBinLabel(loader);
 
 	// Table 1 helper
 	RegisterTableOne(loader);